+++ src/app/card-header/card-header.component.ts
@@ -1,0 +1,16 @@
+import { Component, input, output } from '@angular/core';
+import { IonFabButton, IonIcon } from "@ionic/angular/standalone";
+
+@Component({
+  selector: 'app-card-header',
+  templateUrl: './card-header.component.html',
+  styleUrls: ['./card-header.component.scss'],
+  standalone: true,
+  imports: [IonFabButton, IonIcon]
+})
+export class CardHeaderComponent {
+
+  title = input('Card Header');
+  addClicked = output<void>();
+  addHidden = input<boolean>(false);
+}