+++ src/app/data/data-manager.ts
@@ -392,7 +392,11 @@ export class DataManager implements WorkerClass {
       pinIndex[art.uid] = art.pin;
 
       if (!art.location_string) {
-        art.location_string = LocationName.Unplaced;
+        if (art.art_type?.includes('Vehicle')) {
+          art.location_string = LocationName.Mobile;
+        } else {
+          art.location_string = LocationName.Unplaced;
+        }
         if (art.pin?.x) {
           art.location_string = undefined; // Its placed with x,y
         }
@@ -864,6 +868,7 @@ export class DataManager implements WorkerClass {
         const timeString = this.getTimeString(event, day);
         event.timeString = timeString.short;
         event.longTimeString = timeString.long;
+        event.event_type.label = event.event_type.label.replace(/,/g, ', ');
         event.distance = distance(coords!, event.gpsCoords);
         event.distanceInfo = formatDistance(event.distance);
         if (match == 'Important') {